+++ app/AI4Green4Students/Controllers/NotesController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Models.Note;
 using Models.Section;
+using Models.Section.Form;
 using Models.Stage;
 using Services;
