+++ app/AI4Green4Students/Controllers/PlansController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Models.Plan;
 using Models.Section;
+using Models.Section.Form;
 using Models.Stage;
 using Services;
