+++ app/AI4Green4Students/Controllers/ProjectGroupsController.cs
@@ -8,6 +8,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Models.ProjectGroup;
 using Models.Section;
+using Models.Section.Form;
 using Services;
 
 [ApiController]
