+++ app/AI4Green4Students/Controllers/ReportsController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Models.Report;
 using Models.Section;
+using Models.Section.Form;
 using Models.Stage;
 using Services;
