+++ app/AI4Green4Students/Services/ProjectGroupService.cs
@@ -13,6 +13,7 @@ using Microsoft.EntityFrameworkCore;
 using Models.Emails;
 using Models.ProjectGroup;
 using Models.Section;
+using Models.Section.Form;
 
 public class ProjectGroupService
 {
